fix(poke): align superuser admin with backend rules
